Disallow users from typing two spaces
@uberbryn I couldn't make it so we don't automatically type the space,
so I've just prevented users from double typing spaces in general.
