+++ src/modifiers/handleInlineStyle.js
@@ -2,6 +2,7 @@ import changeCurrentInlineStyle from "./changeCurrentInlineStyle";
 import { EditorState, Modifier } from "draft-js";
 import { inlineMatchers } from "../constants";
 import insertText from "./insertText";
+import { getCurrentLine as getLine } from "../utils";
 
 const handleChange = (editorState, line, whitelist) => {
   let newEditorState = editorState;
@@ -27,17 +28,6 @@ const handleChange = (editorState, line, whitelist) => {
   return newEditorState;
 };
 
-const getLine = (editorState, anchorOffset) => {
-  const selection = editorState.getSelection().merge({ anchorOffset });
-  const key = editorState.getSelection().getStartKey();
-
-  return editorState
-    .getCurrentContent()
-    .getBlockForKey(key)
-    .getText()
-    .slice(0, selection.getFocusOffset());
-};
-
 const handleInlineStyle = (
   whitelist,
   editorStateWithoutCharacter,
@@ -45,14 +35,14 @@ const handleInlineStyle = (
 ) => {
   const editorState = insertText(editorStateWithoutCharacter, character);
   let selection = editorState.getSelection();
-  let line = getLine(editorState, selection.getAnchorOffset());
+  let line = getLine(editorState);
   let newEditorState = handleChange(editorState, line, whitelist);
   let lastEditorState = editorState;
 
   // Recursively resolve markdown, e.g. _*text*_ should turn into both italic and bold
   while (newEditorState !== lastEditorState) {
     lastEditorState = newEditorState;
-    line = getLine(newEditorState, selection.getAnchorOffset());
+    line = getLine(newEditorState);
     newEditorState = handleChange(newEditorState, line, whitelist);
   }
