+++ src/modifiers/insertText.js
@@ -3,6 +3,7 @@ import { EditorState, Modifier } from "draft-js";
 const insertText = (editorState, text) => {
   const selection = editorState.getSelection();
   const content = editorState.getCurrentContent();
+  if (!selection.isCollapsed()) return editorState;
   const newContentState = Modifier.insertText(
     content,
     selection,
